### 2021-02-04 #3008 @robashaw
> @robashaw if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Possible reviewers taken from the list, based on being comfortable with C++ and having expertise in computational chemistry/electronic structure:

lorenzo-rovigatti
felipeZ
nicorenaud
awgwvk
smutao
xwang862
