### 2021-02-04 #3008 @poulson
Hi Lorenzo (@lorenzo-rovigatti) :wave:.

In addition to you being recommended by Robert Shaw (@robashaw) as an expert reviewer for this submission, your C++/Python project BaggieAnalysis also seems to be quite relevant (https://github.com/lorenzo-rovigatti/baggianalysis).

Please let me know if you would be willing and able to review this submission on "libecpint: A C++ library for the eﬀicient evaluation of integrals over effective core potentials".
