### 2021-02-04 #3008 @poulson
Hi Felipe (@felipeZ) :wave:.

In addition to your work on the C++ VOCTA-XTP library (https://github.com/votca/xtp/commits?author=felipeZ), you have been recommended as an expert by Robert Shaw (@robashaw).

Please let me know if you would be willing and able to review this submission on "libecpint: A C++ library for the eﬀicient evaluation of integrals over effective core potentials".
