### 2021-02-05 a2274a4
cmake: update version in project()
The version in project() is used by cmake when exporting the
targets.
