### 2021-02-06 f6f613f
cmake: update version in project() (#32)
The version in project() is used by cmake when exporting the
targets.
