### 2021-02-06 0cca750
Update cmake/make_release.sh
