### 2021-02-10 d53a409
add make_release script (#33)
* add make_release script

* Update cmake/make_release.sh
