### 2021-02-15 #3039 @whedon
[REVIEW]: libecpint: A C++ library for the eﬀicient evaluation of integrals over effective core potentials
**Submitting author:** @robashaw (<a href="http://orcid.org/0000-0002-9977-0835">Robert A. Shaw</a>)
**Repository:** <a href="https://github.com/robashaw/libecpint" target ="_blank">https://github.com/robashaw/libecpint</a>
**Version:** v1.0.4
