### 2021-02-15 #3039 @felipeZ
Dear Robert (@robashaw), I think that you library is a much needed contribution for the development of open source software on computational chemistry. I also think that the API is straightforward and the documentation is pretty clear.
If you agree, I will open some issues in the [libecpint repository](https://github.com/robashaw/libecpint) that I will link back here to discuss several topics about the software that are still not clear to me.
