### 2021-02-15 robashaw/libecpint#34 @felipeZ
Contributing to libecpint
Firstly, thanks for your time and effort developing this great library.
As part of the [libecpint's JOSS review](https://github.com/openjournals/joss-reviews/issues/3039) I would like to add  some thoughts about making it easier to contribute to your code.

## Some ideas about contributing
* In the repository I am missing a *Code of Conduct*. Have  a look to the [open source code of conduct](https://opensource.guide/code-of-conduct/)
* As a developer, the most time consuming aspect when contributing to a new project is to understand the general architecture of the software, thinks like where are the main classes, how are the templates used, what are the conventions, etc. Though the API documentation greatly help to navigate the library, I think that having some informal but concrete description of the architecture will help significantly help new pontential developer. See [this interesting article](https://matklad.github.io//2021/02/06/ARCHITECTURE.md.html) about adding an `ARCHITECTURE.md` file
