### 2021-02-15 robashaw/libecpint#35 @felipeZ
Clean and document auxiliar scripts
As part of the [libecpint's JOSS review](https://github.com/openjournals/joss-reviews/issues/3039) I would suggest to add comments and it possible to improve the code quality of the following miscellaneous scripts:
* [parseecp](https://github.com/robashaw/libecpint/blob/master/share/libecpint/parseecp.py)
* [unrol_radial](https://github.com/robashaw/libecpint/blob/master/src/generated/radial/unrol_radial.py)

Although these scripts are not part of the core functionality, they parse data that is essential for the library. So, It would be great to have some minimal documentation of what they do.
