### 2021-02-15 #3039 @lorenzo-rovigatti
Hi Robert (@robashaw), thanks a lot for your contribution. The libecpint library looks neat!
I'll also be opening a few issues in the [libecpint repo](https://github.com/robashaw/libecpint) in the next days to discuss some of the aspects that may need improvement.
