### 2021-02-16 robashaw/libecpint#36 @lorenzo-rovigatti
Prepare commented real-world examples
As part of the [libecpint's JOSS review](https://github.com/openjournals/joss-reviews/issues/3039) I suggest to prepare a couple of examples that solve some (possibly simple) real world problems.

I appreciate that there are already a few high-level tests that users can adapt to write their own code, but they are

1. a bit buried in the source tree;
2. not commented (that is, it's not clear what they do);
3. they are compiled & executed as tests and therefore are not standalone (from a new user's perspective)

In my opinion the simplest thing you could do would be to take one or both high-level tests that are already present (`api_test1` and `api_test2`), comment them lightly (or write a short accompanying README file) and put them in a separate `examples` folder (to be placed in the root folder perhaps?). There should also be a self-contained CMakeLists.txt that shows how to compile a project that uses the lib once it is installed.
