### 2021-02-16 robashaw/libecpint#37 @lorenzo-rovigatti
Default compilation options
This is an issue created as part of the [libecpint's JOSS review](https://github.com/openjournals/joss-reviews/issues/3039).

I have noticed that, by default, libecpint is compiled without any optimisations. What's the rationale behind this choice? In most scientific libraries/software I have used the default behaviour is usually to compile with some (if not maximum) optimisation, especially if, like in this case, the library can be used as a performance-critical kernel.

Since you use cmake you can use a portable solution by letting the user choose the value of the [`CMAKE_BUILD_TYPE`](https://cmake.org/cmake/help/latest/variable/CMAKE_BUILD_TYPE.html) variable, which could be set to `Release` by default but could be changed to `Debug` or `RelWithDebInfo` when debugging. Of course, further options could still be passed with `CMAKE_CXX_FLAGS`.
