### 2021-03-04 8d69d20
Added default CMake Build Type
