+++ share/libecpint/parseecp.py
@@ -1,9 +1,27 @@
+# Script to parse ECP files into XML format for use by libecpint
+# Usage: python parseecp.py [NAME]
+# will convert raw/NAME.ecp into xml/NAME.xml
+# which can then be found by libecpint using reference "NAME"
 from lxml import etree
 
+# letter symbols for angular momentum shells
+# can be added to if needed
 angular_shells = ["s", "p", "d", "f", "g", "h", "i"]
 
+
 class Shell:
-    def __init__(self, lval = 0, nexp = 0, nbfs = 0):
+    """ Container for a shell of an ECP
+        i.e. the Gaussian expansion of a fixed angular momentum
+        
+        Members:
+        lval - angular momentum quantumm number of shell
+        powers - the power of r for each Gaussian in the shell
+        nexp - the number of primitive Gaussians (exponents) in the shell
+        exps - the exponents of the Gaussians
+        contr - the contraction coefficients of the Gaussians
+    """
+    def __init__(self, lval = 0, nexp = 0):
+        """Creates a  Shell with angular momentum lval and nexp exponents"""
         self.lval = lval
         self.powers = []
         self.nexp = nexp
@@ -11,50 +29,83 @@ class Shell:
         self.contr = []
 
 class Atom:
+    """ Container for an ECP for a specific atom type
+        
+        Members:
+        name - name of the atom, e.g. 'O' for oxygen
+        ncore - the number of core electrons in the ECP
+        maxl - the maximum angular momentum shell in the ECP
+        nshells - the number of shells in the ECP
+        shells - an array of Shell objects describing the shells in the ECP
+    """
     def __init__(self, name="X", nshells = 0):
+        """Creates a blank Atom with the given name"""
         self.name = name
         self.ncore = 0
         self.maxl = 0
         self.nshells = nshells
         self.shells = []
         
+def tokenize(line, sep=','):
+    """Given a line of input, cleans up and returns tokens,
+       split by the separator (sep)
+    """
+    # strip out whitespace and split by comma
+    line = lines[linenumber].strip()
+    tokens = line.split(sep)
+    # get rid of additional whitespace
+    for token in tokens: 
+        token = token.replace(' ', '')
+    return tokens
+        
 def parse_ecp(file):
+    """Given a MOLPRO-format ECP file, returns Atom objects (ECPs) for every
+       atom type defined in that file.  
+    """
     atoms = []
     atomnames = {}
+    
+    # read in the file
     lines = file.readlines()
     nlines = len(lines)
+    
+    # loop to the end of the file
     linenumber = 0
     atomnumber = 0
-    
     while linenumber < nlines:
-        line = lines[linenumber].strip()
-        tokens = line.split(',')
-        for token in tokens: 
-            token = token.replace(' ', '')
-            
+        tokens = tokenize(lines[linenumber], sep=',')
+        
+        # all lines where ECP definitions start will have 
+        # at least two bits "ecp,AtomName"
         if (len(tokens) > 2):
             if (tokens[0].lower() == "ecp"):
+                # Found an ECP definition
                 atom_name = tokens[1].lower()
+                # Create a container for the ECP
                 new_atom = Atom(name=atom_name)
 
+                # the next two tokens should be the no. of core electrons
+                # and the maximum angular momentum shell of the ECP
                 new_atom.ncore = int(tokens[2])
                 new_atom.maxl = int(tokens[3])
-                
+                # there should then be maxl+1 lines definining the shells
+                # in the order [maxL, 0, 1, ..., maxL-1]                
                 for i in range(new_atom.maxl+1):
                     linenumber += 1
-                    line = lines[linenumber].strip()
+                    tokens = tokenize(lines[linenumber], sep=';')
                     
-                    tokens = line.split(';')
-                    for token in tokens:
[…]
