### 2021-03-06 78a7777
Added an ARCHITECTURE file
