### 2021-03-06 robashaw/libecpint#37 @robashaw
Thank you for the suggestion - I'm honestly not sure why it wasn't there already. The default build is now Release with optimizations turned on, and will be added in the linked pull request.
