### 2021-03-06 robashaw/libecpint#36 @robashaw
An example use with instructions on how to build has been added as part of the linked PR.
