### 2021-03-06 robashaw/libecpint#35 @robashaw
This has now been done as part of the linked PR.
