### 2021-03-06 robashaw/libecpint#34 @robashaw
Thank you. The article was interesting to read, and I agree that is one of the hardest parts of joining a new project. As such, I've written an architecture overview, and added the Contributor Covenant code of conduct, in the linked PR.
