### 2021-03-06 828c309
Changed to std::copy in example
