### 2021-03-08 robashaw/libecpint#36 @lorenzo-rovigatti
Thank you! Just a warning: the suggested way of compiling the example does not work on my machine (where I tried with both `zsh` and `bash`). I had to remove the square brackets, that is, change from

```
cmake .. [-DCMAKE_CXX_FLAGS="-I/home/lorenzo/installazioni/include -std=c++14"] [-DCMAKE_EXE_LINKER_FLAGS="-L/home/lorenzo/installazioni/lib"]
```

to

```
cmake .. -DCMAKE_CXX_FLAGS="-I<include_dir> -std=c++14"] -DCMAKE_EXE_LINKER_FLAGS="-L<lib_dir>"
```
