### 2021-03-08 a7a3d88
Updated Example/Readme to get rid of square brackets
