+++ example/README.md
@@ -7,7 +7,7 @@ This directory contains a working example of how to use libecpint for its core f
 To build this, make sure you have built libecpint (and its dependencies, pugixml and Faddeeva). Then using cmake in this directory
 
 ```
-cmake . -Bbuild [-DCMAKE_CXX_FLAGS="-I/include/dirs -std=c++14"] [-DCMAKE_EXE_LINKER_FLAGS="-L/lib/dirs"] 
+cmake . -Bbuild -DCMAKE_CXX_FLAGS="-I/include/dirs -std=c++14" -DCMAKE_EXE_LINKER_FLAGS="-L/lib/dirs"
 ```
 where the -I and -L flags are used to point to the libecpint headers and library, respectively, if these are not already in your path.
 
@@ -15,7 +15,7 @@ To then build the library and run the test:
 ```
 cd build
 make
-./example [LIBECPINT_SHARE_DIR]
+./example LIBECPINT_SHARE_DIR
 ```
 where the argument points to the share/libecpint directory.
 
