### 2021-03-19 #3039 @robashaw
@poulson @felipeZ @lorenzo-rovigatti If there are no further changes, I will merge [https://github.com/robashaw/libecpint/pull/38](this libecpint pull request). Thank you for your time and helpful suggestions :)
