### 2021-03-22 f40bbc2
JOSS review changes (#38)
* Started JOSS paper

* Added references to JOSS paper

* Added workflow to build JOSS paper

* Testing GH action

* Fixed yml placement

* Added missing reference to paper.bib

* Fixed formatting issues with references

* Fixed typos in paper.md

* Fixed Serenity reference

* Added default CMake Build Type

* Added comments to parseecp

* Fixed typo in parseecp

* Documented unrol_radial

* Added Code of Conduct

* Added working example (HI/VTZ-PP)

* Added an ARCHITECTURE file

* Changed to std::copy in example

* Updated Example/Readme to get rid of square brackets
[…]
