+++ doc/joss/paper.bib
@@ -1,0 +1,216 @@
+@incollection{Dolg2000,
+	edition = {Second},
+	series = {{NIC} series},
+	title = {Effective {Core} {Potentials}},
+	volume = {3},
+	isbn = {3-00-005834-6},
+	booktitle = {Modern {Methods} and {Algorithms} in {Quantum} {Chemistry}},
+	publisher = {John von Neumann Institute for Computing},
+	author = {Dolg, M},
+	editor = {Grotendorst, Johannes},
+	year = {2000},
+	pages = {507--540}
+}
+
+@incollection{Dolg2002,
+	title = {Chapter 14 {Relativistic} effective core potentials},
+	volume = {11},
+	isbn = {978-0-444-51249-9},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S1380732302800401},
+	urldate = {2021-01-28},
+	booktitle = {Theoretical and {Computational} {Chemistry}},
+	publisher = {Elsevier},
+	author = {Dolg, M.},
+	year = {2002},
+	doi = {10.1016/S1380-7323(02)80040-1},
+	pages = {793--862}
+}
+
+@article{Wadt1985,
+	title = {\textit{{Ab} initio} effective core potentials for molecular calculations. {Potentials} for main group elements {Na} to {Bi}},
+	volume = {82},
+	issn = {0021-9606, 1089-7690},
+	url = {http://aip.scitation.org/doi/10.1063/1.448800},
+	doi = {10.1063/1.448800},
+	number = {1},
+	urldate = {2021-01-28},
+	journal = {The Journal of Chemical Physics},
+	author = {Wadt, Willard R. and Hay, P. Jeffrey},
+	month = jan,
+	year = {1985},
+	pages = {284--298}
+}
+
+@article{Hay1985,
+	title = {\textit{{Ab} initio} effective core potentials for molecular calculations. {Potentials} for the transition metal atoms {Sc} to {Hg}},
+	volume = {82},
+	issn = {0021-9606, 1089-7690},
+	url = {http://aip.scitation.org/doi/10.1063/1.448799},
+	doi = {10.1063/1.448799},
+	number = {1},
+	urldate = {2021-01-28},
+	journal = {The Journal of Chemical Physics},
+	author = {Hay, P. Jeffrey and Wadt, Willard R.},
+	month = jan,
+	year = {1985},
+	pages = {270--283}
+}
+
+@article{Dolg2012,
+	title = {Relativistic {Pseudopotentials}: {Their} {Development} and {Scope} of {Applications}},
+	volume = {112},
+	issn = {0009-2665, 1520-6890},
+	shorttitle = {Relativistic {Pseudopotentials}},
+	url = {https://pubs.acs.org/doi/10.1021/cr2001383},
+	doi = {10.1021/cr2001383},
+	number = {1},
+	urldate = {2021-01-28},
+	journal = {Chemical Reviews},
+	author = {Dolg, M. and Cao, X.},
+	month = jan,
+	year = {2012},
+	pages = {403--480}
+}
+
+@article{McMurchie1981,
+	title = {Calculation of integrals over ab initio pseudopotentials},
+	volume = {44},
+	issn = {00219991},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/002199918190053X},
+	doi = {10.1016/0021-9991(81)90053-X},
+	number = {2},
+	urldate = {2021-01-28},
+	journal = {Journal of Computational Physics},
+	author = {McMurchie, Larry E and Davidson, Ernest R},
+	month = dec,
+	year = {1981},
+	pages = {289--301},
[…]
