+++ README.md
@@ -1,4 +1,4 @@
-## Libecpint 1.0.5
+# Libecpint 1.0.5
 
 [![Build Status](https://dev.azure.com/robertshaw383/libecpint/_apis/build/status/robashaw.libecpint?branchName=master)](https://dev.azure.com/robertshaw383/libecpint/_build/latest?definitionId=2&branchName=master)
 [![codecov](https://codecov.io/gh/robashaw/libecpint/branch/master/graph/badge.svg)](https://codecov.io/gh/robashaw/libecpint)
@@ -6,6 +6,10 @@
 [![Code Quality](https://www.code-inspector.com/project/15206/status/svg)]()
 
 Libecpint is a C++ library for the efficient evaluation of integrals over ab initio effective core potentials, using a mixture of generated, recursive code and Gauss-Chebyshev quadrature. It is designed to be standalone and generic, and is now in its first stable release. If you experience any problems please raise an issue here; contributions and suggestions are also welcome.
+
+## Contributing
+
+Contributions are welcomed, either in the form of raising issues or pull requests on this repo. Please take a look at the Code of Conduct before interacting, which includes instructions for reporting any violations.
 
 ## New in first full release
 
@@ -49,6 +53,10 @@ Additionally, if you wish to regenerate the radial code (see below),  Python >=3
 ## Documentation
 
 Please refer to the main documentation [here](https://libecpint.readthedocs.io/en/latest/index.html).
+
+## Examples
+
+There is also a working example in the example folder, with instructions of how to build and link against the library. Please also the API tests in tests/lib/
 
 ## Acknowledging usage
 
