+++ include/generate.hpp
@@ -125,7 +125,7 @@ namespace libecpint {
 			int f6 = CA == s.CA ? 1 : 0;
 			int f7 = CB == s.CB ? 1 : 0; 
 		
-			return {f1, f2, f3, f4, f5, f6, f7};  
+			return Heptuple{f1, f2, f3, f4, f5, f6, f7};  
 		}
 	
 		/// Prints out a SumTerm without compressing the indices - currently preferred 
