### 2021-04-02 e0c29b6
Optionally compile without pugixml and disable ECP definition file reading
