### 2021-04-03 6d6a041
Hide standard out printing behind preprocessor definition
