### 2021-04-04 46e9280
Allow compiling without pugixml (#41)
* Ignore output from compile_commands.json and language server caches

* Optionally compile without pugixml and disable ECP definition file reading

* Formatting for CMake status messages

* Hide standard out printing behind preprocessor definition

Co-authored-by: Robert Shaw <robertshaw383@gmail.com>
