### 2021-04-15 8fa5119
Should not compile without pugixml (#43)
