### 2021-04-15 4ad71a8
Make C++ standard more explicit (#44)
