+++ CMakeLists.txt
@@ -1,6 +1,11 @@
 # minimum version of CMake is 3.12
 cmake_minimum_required(VERSION 3.12 FATAL_ERROR)
 
+include(CheckCXXCompilerFlag)
+set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_EXTENSIONS OFF) #...without compiler extensions like gnu++11
+
 #build type
 if(NOT CMAKE_BUILD_TYPE)
   message(STATUS "No build type specified. Will build Release")
