### 2021-04-15 7284ebd
Bump version to 1.0.6
