+++ CMakeLists.txt
@@ -19,7 +19,7 @@ endif(NOT CMAKE_BUILD_TYPE)
 
 # define project
 project( libecpint
-         VERSION 1.0.5
+         VERSION 1.0.6
          LANGUAGES C CXX)
 
 set(API_VERSION 1)
