### 2021-04-15 ed8bebd
Update to release v1.0.6
